Further API Design
- Own error type
- traits for Role and Policy
- trait for a SSOConnector

The SSOConnector trait requires the used types for role and policy to
implement the corresponding traits. This enables the user to use his own
types for roles and policies.

There will be default implementations for keycloak, authelia and CAS but
any SSO service can be easily extended by implementing the traits.